Fixed test message and file creation unit
